connections are now handled by a connection manager. Connections have a list of streams. When the user passes a packet to the protocol object, the object returns the corresponding connection or creates one if it doesn't exist, but doesn't actually process the packet. The user can then pass the packet to the connection object whenever it wants, or not at all

+++ StreamUPLib/StreamUP/Data/include/Data/buffer/ReadBufferWrapper.h
@@ -10,7 +10,7 @@ namespace SUP::BufferUtil
 class ReadBufferWrapper
 {
 public:
-    ReadBufferWrapper(const uint8_t *data, int length) : data(data), length(length), readIndex(0) {}
+    ReadBufferWrapper(const uint8_t *data, size_t length) : data(data), length(length), readIndex(0) {}
 
     template<typename T>
     void read(T &value)
@@ -25,23 +25,23 @@ public:
         return tmp;
     }
 
-    void readToArray(uint8_t *dest, int numBytes)
+    void readToArray(uint8_t *dest, size_t numBytes)
     {
         std::copy_n(data + readIndex, numBytes, dest);
         readIndex += numBytes;
     }
 
-    void setReadIndex(int index) { readIndex = index; }
+    void setReadIndex(size_t index) { readIndex = index; }
 
-    [[nodiscard]] int getReadIndex() const { return readIndex; }
+    [[nodiscard]] size_t getReadIndex() const { return readIndex; }
 
-    [[nodiscard]] int getReadableBytes() const { return length - readIndex; }
+    [[nodiscard]] size_t getReadableBytes() const { return length - readIndex; }
 
-    [[nodiscard]] bool ensureReadableBytes(int numBytes) const { return getReadableBytes() >= numBytes; }
+    [[nodiscard]] bool ensureReadableBytes(size_t numBytes) const { return getReadableBytes() >= numBytes; }
 
 private:
     const uint8_t *data;
-    int length;
-    int readIndex;
+    size_t length;
+    size_t readIndex;
 };
 }
